// Paylocity.CodingChallenge.UI/src/models/EmployeeCosts.ts
export class EmployeeCosts  {
    totalCostOfBenefitsPerYear: number = 0;
    salaryPerPayCheck: number = 0;
    salaryAfterBenefitsDeductionPerPayCheck: number = 0;

    name: string;
  dependentCount: number = 0;
  annualSalary: number = 0;
  monthlySalary: number = 0;
  totalDeductions: number = 0;
  finalAnnualSalary: number = 0;
  finalMonthlySalary: number = 0;
  payChecksPerYear: number = 26
  }